riscv-page-tables: move test next to code tested.
Pull the shared test code to its own file so it can be used from
multiple places. The shared code was the only reason the tests for the
page table modules were bundled in lib.rs instead of local to where the
page tables are implemented.
